Added a base class fro all *Mapper classes of the Music app
- The *Mapper classes used to be inherited from the core class OCP\AppFramework\Db\Mapper, now there is an additional level OCA\Music\Db\BaseMapper which inherits the core class and acts as base class for the other *Mapper classes
- The base class implements the methods 'deleteById' and 'count' which used to have identical implementations copy-pasted in several *Mapper classes
- This should make Scrutinizer happy
