+++ db/albummapper.php
@@ -13,10 +13,9 @@
 namespace OCA\Music\Db;
 
 use OCP\AppFramework\Db\DoesNotExistException;
-use OCP\AppFramework\Db\Mapper;
 use OCP\IDb;
 
-class AlbumMapper extends Mapper {
+class AlbumMapper extends BaseMapper {
 
 	public function __construct(IDb $db){
 		parent::__construct($db, 'music_albums', '\OCA\Music\Db\Album');
@@ -190,22 +189,6 @@ class AlbumMapper extends Mapper {
 	}
 
 	/**
-	 * @param integer[] $albumIds
-	 */
-	public function deleteById($albumIds){
-		if(count($albumIds) === 0) {
-			return;
-		}
-
-		$questionMarks = array();
-		for($i = 0; $i < count($albumIds); $i++){
-			$questionMarks[] = '?';
-		}
-		$sql = 'DELETE FROM `*PREFIX*music_albums` WHERE `id` IN ('. implode(',', $questionMarks) . ')';
-		$this->execute($sql, $albumIds);
-	}
-
-	/**
 	 * @param integer $coverFileId
 	 * @param integer $parentFolderId
 	 */
@@ -289,19 +272,6 @@ class AlbumMapper extends Mapper {
 				SET `cover_file_id` = ? WHERE `id` = ?';
 		$params = array($imageId, $albumId);
 		$this->execute($sql, $params);
-	}
-
-	/**
-	 * @param string $userId
-	 * @return integer
-	 */
-	public function count($userId){
-		$sql = 'SELECT COUNT(*) AS count FROM `*PREFIX*music_albums` '.
-			'WHERE `user_id` = ?';
-		$params = array($userId);
-		$result = $this->execute($sql, $params);
-		$row = $result->fetch();
-		return $row['count'];
 	}
 
 	/**
